Move the rest method into FastHttpUser instead of RestUser. Remove RestUser. Update docs accordingly.
