+++ locust/contrib/fasthttp.py
@@ -4,10 +4,13 @@ import socket
 import json
 import json as unshadowed_json  # some methods take a named parameter called json
 from base64 import b64encode
+from contextlib import contextmanager
+from json.decoder import JSONDecodeError
 from urllib.parse import urlparse, urlunparse
 from ssl import SSLError
 import time
-from typing import Callable, Optional, Tuple, Dict, Any
+import traceback
+from typing import Callable, Optional, Tuple, Dict, Any, Generator, cast
 
 from http.cookiejar import CookieJar
 
@@ -323,6 +326,8 @@ class FastHttpUser(User):
     abstract = True
     """Dont register this as a User class that can be run by itself"""
 
+    _callstack_regex = re.compile(r'  File "(\/.[^"]*)", line (\d*),(.*)')
+
     def __init__(self, environment):
         super().__init__(environment)
         if self.host is None:
@@ -350,6 +355,56 @@ class FastHttpUser(User):
         Instance of HttpSession that is created upon instantiation of User.
         The client support cookies, and therefore keeps the session between HTTP requests.
         """
+
+    @contextmanager
+    def rest(
+        self, method, url, headers: Optional[dict] = None, **kwargs
+    ) -> Generator[RestResponseContextManager, None, None]:
+        headers = {"Content-Type": "application/json", "Accept": "application/json"} if headers is None else headers
+        with self.client.request(method, url, catch_response=True, headers=headers, **kwargs) as r:
+            resp = cast(RestResponseContextManager, r)
+            resp.js = None  # type: ignore
+            if resp.text is None:
+                # round the response time to nearest second to improve error grouping
+                response_time = round(resp.request_meta["response_time"] / 1000, 1)
+                resp.failure(
+                    f"response body None, error {resp.error}, response code {resp.status_code}, response time ~{response_time}s"
+                )
+            else:
+                if resp.text:
+                    try:
+                        resp.js = resp.json()
+                    except JSONDecodeError as e:
+                        resp.failure(
+                            f"Could not parse response as JSON. {resp.text[:250]}, response code {resp.status_code}, error {e}"
+                        )
+            try:
+                yield resp
+            except AssertionError as e:
+                if e.args:
+                    if e.args[0].endswith(","):
+                        short_resp = resp.text[:200] if resp.text else resp.text
+                        resp.failure(f"{e.args[0][:-1]}, response was {short_resp}")
+                    else:
+                        resp.failure(e.args[0])
+                else:
+                    resp.failure("Assertion failed")
+
+            except Exception as e:
+                error_lines = []
+                for l in traceback.format_exc().split("\n"):
+                    m = self._callstack_regex.match(l)
+                    if m:
+                        filename = re.sub(r"/(home|Users/\w*)/", "~/", m.group(1))
+                        error_lines.append(filename + ":" + m.group(2) + m.group(3))
+                    short_resp = resp.text[:200] if resp.text else resp.text
+                    resp.failure(f"{e.__class__.__name__}: {e} at {', '.join(error_lines)}. Response was {short_resp}")
+
+    # some web api:s use a timestamp as part of their url (to break thru caches). this is a convenience method for that.
+    @contextmanager
+    def rest_(self, method, url, name=None, **kwargs) -> Generator[RestResponseContextManager, None, None]:
+        with self.rest(method, f"{url}&_={int(time.time()*1000)}", name=name, **kwargs) as resp:
+            yield resp
 
 
 class FastRequest(CompatRequest):
@@ -572,3 +627,8 @@ class ResponseContextManager(FastResponse):
         if not isinstance(exc, Exception):
             exc = CatchResponseError(exc)
         self._manual_result = exc
+
+
+class RestResponseContextManager(ResponseContextManager):
+    js: dict  # This is technically an Optional, but I dont want to force everyone to check it
+    error: Exception  # This one too
