+++ tox.ini
@@ -28,7 +28,7 @@ allowlist_externals =
 commands =
     python3 -m pip install .
     python3 -m coverage run -m unittest discover []
-    bash -ec "PYTHONUNBUFFERED=1 timeout 10s python3 examples/rest_ex.py >out.txt 2>err.txt || true"
+    bash -ec "PYTHONUNBUFFERED=1 timeout 10s python3 examples/rest.py >out.txt 2>err.txt || true"
     grep -qm 1 'my custom error message with response text, response was {"args"' out.txt
     grep -qm 1 'ZeroDivisionError: division by zero at.*Response was {"ar' out.txt
     bash -ec '! grep . err.txt' # should be empty
